Thought of better way to do this. Use the app array to find the tables that use the v_id then delete the tenant settings from the tables that do use v_id.
git-svn-id: http://fusionpbx.googlecode.com/svn/trunk@1837 e2954ef0-caed-11de-bd98-e341d93cc3cc

+++ fusionpbx/core/system_settings/v_system_settings_delete.php
@@ -86,16 +86,11 @@ else {
 				foreach ($apps as &$app) {
 					foreach ($app['db'] as $row) {
 						$table_name = $row['table'];
-						switch ($table_name) {
-							case "v_menus":
-								break;
-							case "v_menu_items":
-								break;
-							case "v_menu_item_groups":
-								break;
-							default:
-							  $sql = "delete from $table_name where v_id = '$id' ";
-							  $db->query($sql);
+						foreach ($row['fields'] as $field) {
+							if ($field['name'] == "v_id") {
+								$sql = "delete from $table_name where v_id = '$id' ";
+								$db->query($sql);
+							}
 						}
 					}
 				}
